web_platform/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==23.2.1
authlib==1.3.0
httpx==0.27.0
itsdangerous==2.2.0
psycopg2-binary==2.9.9
